tabs added to propProfile page
