+++ src/pages/building-summary-components/floorPlanAvgLease.tsx
