+++ src/pages/building-summary-components/floorPlanCount.tsx
