+++ src/pages/building-summary-components/leaseTrends.tsx
