+++ src/pages/building-summary-components/lossToLease.tsx
