+++ src/pages/building-summary-components/recentOccupancy.tsx
