+++ src/pages/building-summary-components/vacancy.tsx
