+++ src/pages/property-profile/floorPlanAnalysis.tsx
@@ -1,0 +1,7 @@
+export function FloorPlanAnalysisComponent() {
+    return(
+        <>
+        Floor Plan Analysis 
+        </>
+    )
+}
