+++ src/pages/property-profile/rawDataChart.tsx
@@ -1,0 +1,7 @@
+export function RawBuildingDataComponent() {
+    return(
+        <>
+        Raw Building Data
+        </>
+    )
+}
